requirement.txt
langchain_core
python-dotenv
streamlit==1.30.0
groq
langchain-groq 